add checker

+++ gofra_service/backend/internal/domain/model_user.go
@@ -11,7 +11,7 @@ type User struct {
 	ID        primitive.ObjectID `bson:"_id,omitempty"` // Уникальный идентификатор
 	Login     string             `bson:"login"`         // Логин пользователя
 	PassHash  []byte             `bson:"pass_hash"`     // Хэш пароля
-	Balance   int32              `bson:"balance"`       // Баланс пользователя в минимальных единицах валюты (например, центы)
+	Balance   int64              `bson:"balance"`       // Баланс пользователя в минимальных единицах валюты (например, центы)
 	CreatedAt time.Time          `bson:"created_at"`    // Время создания
 	// Добавить вывод в профиле пользователя информации о выставленных лотах
 }
